tests: move to a better place
